[Bronze III] Title: 전자레인지, Time: 76 ms, Memory: 11468 KB -BaekjoonHub

+++ 백준/Bronze/10162. 전자레인지/전자레인지.java
@@ -5,7 +5,7 @@ class Main {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 
         // 1. 입력 받기
-        int t = Integer.parseInt(br.readLine());
+        long t = Integer.parseInt(br.readLine());
         int[] times = {3000, 60, 10};
 
         // 2. T초를 맞출 수 없을 때
